Silent notifications.
The application will add a notification whenever a new user is added to the queue.
Clicking on it sends the user to the login screen, due to what appears to be a
limitation in how the application is created.  Furthermore, because this is
not done in a service, I do not beleive it will work when the application is
in the background.